requirements.txt
Django==3.2.4
djangorestframework==3.12.4
drf-yasg==1.20.0
mysql-connector-python==8.0.25
flake8==3.9.2

